% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_to_daily.R
\name{monthly_to_daily}
\alias{monthly_to_daily}
\title{Estimate the year's daily precipitation from average monthly values.}
\usage{
monthly_to_daily(p, year = "2020")
}
\arguments{
\item{p}{numeric vector of length 12 corresponding to average millimeters of precipitation for each month.}

\item{year}{character vector of the year}
}
\value{
Vector of length 365 or 366 with daily precipitation data.
}
\description{
Gets a vector of length 365 (or 366 if the selected year is leap-year) with daily precipitation data and a year.
}
\examples{
# example 1:
monthly_to_daily(p = c(4, 10, 12, 14, 17, 22, 10, 24, 20, 12, 11, 6), year = "2020")
# example 2:
rhq(p = monthly_to_daily(p = c(4, 10, 12, 14, 17, 22, 10, 24, 20, 12, 11, 6),
                         year = "2020"),
                         area = 100)

}
