% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palembang_precipitation.R
\docType{data}
\name{palembang_precipitation}
\alias{palembang_precipitation}
\title{NOAA precipitation data for Palembang City, ID}
\format{
data.frame with precipitation data in mm.
}
\source{
\href{https://www.ncdc.noaa.gov/cdo-web/search}{NOAA FTP Archive}
}
\usage{
data(palembang_precipitation)
}
\description{
Data from "Bandar Idara Sultan Mahmud Dadaruddin II" airport weather station in Palembang City, ID.
}
\examples{
data(palembang_precipitation)
}
\references{
NOAA Climate Data Online, 2020
(\href{https://www.ncdc.noaa.gov/cdo-web/search}{NOAA})
}
\keyword{datasets}
\keyword{precipitation}
