% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhp.R
\name{rhp}
\alias{rhp}
\title{Rainwater harvesting preheating. Calculate a rain harvesting system performance using daily rainfall data and n amount of "preheat" years.}
\usage{
rhp(
  p,
  area,
  holidays = 1,
  tank.size = 450,
  users = 4,
  use = 50,
  filter.coefficient = 0.9,
  keep.options = FALSE,
  n = 3
)
}
\arguments{
\item{p}{numeric vector with daily precipitation in milimeters.}

\item{area}{numeric vector of the rainwater catchment area in m2.}

\item{holidays}{numeric vector with water consumption (1) such as a weekday in an office, or days without significant water consumption (0), such as a weekend in an office.}

\item{tank.size}{numeric vector with the water storage capacity in liters.}

\item{users}{numeric vector with the amount of persons using water.}

\item{use}{numeric vector with the daily per capita water demand in liters.}

\item{filter.coefficient}{numeric vector with the fraction of the water that can be used after the filtering process (default 0.90). Typical values are 0.90 for personal hygiene or drinking water and 0.95 for other uses such as WC or irrigation.).}

\item{keep.options}{logical vector specifying if the tank size should be stored as a variable in the output (default FALSE).}

\item{n}{number of iterations (default 3).}
}
\value{
Dataframe (same number of rows as lenght of p) with daily estimates of precipitation, activity, water demand, rainwater supply, used rainwater, tank level, and wasted water.
}
\description{
Runs the function rhq `n` amount of times. In each iteration, the tank level `tlevel` of the last day of the year is passed as an initial tank level `initial.level` for the first day of the following year.
}
\examples{
# Rainwater harvesting estimation performance values for a rainwater harvesting system,
# considering a daily precipitation of 10 mm and a catchment area of 100 m2.
rhp(p = rep(10, 365), area = 100)
# Repeat the calculations but this time with 4 iterations.
rhp(p = rep(10, 365), area = 100, n = 4)

}
