% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhq.R
\name{rhq}
\alias{rhq}
\title{Rainwater harvesting quantification. Calculate a rain harvesting system performance using daily rainfall data.}
\usage{
rhq(
  p,
  area,
  holidays = 1,
  tank.size = 450,
  users = 4,
  use = 50,
  filter.coefficient = 0.9,
  initial.level = 0,
  keep.options = FALSE
)
}
\arguments{
\item{p}{numeric vector with daily precipitation in milimeters.}

\item{area}{numeric vector of the rainwater catchment area in m2.}

\item{holidays}{numeric vector with water consumption (1) such as a weekday in an office, or days without significant water consumption (0), such as a weekend in an office.}

\item{tank.size}{numeric vector with the water storage capacity in liters.}

\item{users}{numeric vector with the amount of persons using water.}

\item{use}{numeric vector with the daily per capita water demand in liters.}

\item{filter.coefficient}{numeric vector with the fraction of the water that can be used after the filtering process (default 0.90). Typical values are 0.90 for personal hygiene or drinking water and 0.95 for other uses such as WC or irrigation.).}

\item{initial.level}{numeric vector of the initial tank water level in liters (default 0).}

\item{keep.options}{logical vector specifying if the tank size should be stored as a variable in the output (default FALSE).}
}
\value{
Dataframe (same number of rows as lenght of p) with daily estimates of precipitation, activity, water demand, rainwater supply, used rainwater, tank level, and wasted water.
}
\description{
Generates a results list of data related to rainwater demand, supply, and use at a rain harvesting system, using
daily rainfall data and other particular inputs parameters.
}
\examples{
# Rainwater harvesting estimation performance values for a rainwater harvesting system,
# considering a daily precipitation of 10 mm and a catchment area of 100 m2.
rhq(p = rep(10, 365), area = 100)
# Repeat the calculations with custom parameters: a tank vlume of 1000L,
# a filter coefficient of 0.95, and
# 3 users with a daily demand of 100L per capita.
rhq(p = rep(10, 365), area = 100,
    holidays = 1, tank.size = 450, users = 4,
    use = 50, filter.coefficient = 0.9)

}
