% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rh.R
\name{summarize_rh}
\alias{summarize_rh}
\title{Summarize rainwater harvesting results}
\usage{
summarize_rh(x, year = "2020")
}
\arguments{
\item{x}{data frame object}

\item{year}{character vector of the year}
}
\value{
List with 1 dataframe (12 rows) of monthly estimates of demanded and used rainwater, and 2 numeric vectors of water demand percentage covered with rain and liters used in the year.
}
\description{
Gets x, the output from rhq or rhp and the year
}
\examples{
rh_performance <- rhp(p = rep(0.5, 366), area = 100)
summarize_rh(rh_performance)

}
